% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmcs_calculator_binaryMode.R
\name{HMCS_calculatorBinary}
\alias{HMCS_calculatorBinary}
\title{HMCS calcualtor for binary datasets (Example: viable tumor vs all)}
\usage{
HMCS_calculatorBinary(DSC_df, focusROI = "VT", ...)
}
\arguments{
\item{DSC_df}{a dataframe, containing the following vectors: mzList, dsc_mpm_VT.}

\item{focusROI}{a character, specifying which tissue histology feaure you want to focus on, for example: "VT".}

\item{...}{ignored.}

\item{regions}{a \code{regions} object, see \code{?getRegions}.}
}
\value{
Calculates HMCS in binary mode, returns HMCS values from DSC.
}
\description{
This functions calculates HMCS values for binary comparisons.
}
\author{
Shad Arif Mohammed, \email{s.mohammed@doktoranden.hs-mannheim.de}
}
