% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOptical2_onHE.R
\name{plotOptical2}
\alias{plotOptical2}
\title{Superimpose MSI pixels on an optical plot}
\usage{
plotOptical2(
  msiTeachingImgPath,
  optical_imgName,
  targetImgPath,
  HE_imgName,
  regions,
  coords,
  transMat,
  InverseTransMat = FALSE,
  outputPath,
  raster = TRUE,
  polygons = FALSE,
  polygons.fill = FALSE,
  col = "black",
  lwd = 2,
  pch = 19,
  point.cex = 0.1,
  alpha = 0.5,
  label.cex = 1,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{msiTeachingImgPath}{a character, path to the optical image used for MSI teaching.}

\item{targetImgPath}{a character, path to the target optical image (e.g., H&E).}

\item{regions}{a \code{regions} object, see \code{?getRegions}.}

\item{coords}{a list of data.frames representing the different classes/regions
to be transfomed, each with 2 columns, x and y.}

\item{transMat}{a 3x3 transformation matrix for projecting points from \code{msiTeachingImg} (moving)
to \code{targetImg} (fixed).}

\item{InverseTransMat}{a logical, whether to use the inverse of \code{transMat}.}

\item{outputPath}{a character, path for writing the resulting images.}

\item{raster}{logical, whether to plot the points specified in \code{coords}.}

\item{polygons}{a logical (ignored for now), whether to compute a plot the boundaries of the point
specified in \code{coords}.}

\item{polygons.fill}{a logical (ignored for now), whether to fill the computed boundaries of the point
specified in \code{coords}.}

\item{col}{character, the color to be used for points/polygons.}

\item{lwd}{a numeric, line width of polygons.}

\item{pch}{an integer, the points type.}

\item{point.cex}{a numeric, the size of the points for the raster.}

\item{alpha}{a fraction, alpha chanel for the raste points.}

\item{label.cex}{a numeric, the size of the region label text.}

\item{xlim}{a numeric vector specifying \code{xlim} in msi coordinates.}

\item{ylim}{a numeric vector specifying \code{ylim} in msi coordinates.}

\item{...}{ignored.}
}
\value{
Generates plots and writes them to \code{outputPath}.
}
\description{
This functions plots MSI measurement locations (pixels) on top of the
corresponding (reference) optical image for visual inspection.
}
\author{
Denis Abu Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
