% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOptical4.R
\name{plotOptical4}
\alias{plotOptical4}
\title{Superimpose MSI pixels on an optical plot}
\usage{
plotOptical4(
  msiTeachingImgPath,
  targetImgPath,
  regions,
  mpm,
  transMat,
  InverseTransMat = FALSE,
  outputPath,
  figLabel = "mz",
  plotPoints = TRUE,
  plotContours = TRUE,
  whichContours = "both",
  smooth = FALSE,
  col.hotspot = "red",
  col.coldspot = "blue",
  lwd = 2,
  pch = 19,
  point.size = 0.1,
  alpha = 0.5,
  label.cex = 1,
  xlim = NULL,
  ylim = NULL,
  colPal = "turbo",
  ...
)
}
\arguments{
\item{msiTeachingImgPath}{a character, path to the optical image used for MSI teaching.}

\item{targetImgPath}{a character, path to the target optical image (e.g., H&E).}

\item{regions}{a \code{regions} object, see \code{?getRegions}.}

\item{mpm}{an \code{mpm} object.}

\item{transMat}{a 3x3 transformation matrix for projecting points from \code{msiTeachingImg} (moving)
to \code{targetImg} (fixed).}

\item{InverseTransMat}{a logical, whether to use the inverse of \code{transMat}.}

\item{outputPath}{a character, path for writing the resulting images.}

\item{plotPoints}{logical, whether to plot the points and their intensities specified in \code{mpm}.}

\item{plotContours}{a logical (ignored for now), whether to plot the hotspot/coldspot contours of the
\code{mpm} object.}

\item{whichContours}{a character, \code{c("hotspot", "coldspot", "both")}.}

\item{smooth}{a logical, whether to fill the hotspot/coldspot contours with a smoothed
version of intensities computed via the KDE.}

\item{col.hotspot}{character, the color to be used for hotspots contours}

\item{col.coldspot}{character, the color to be used for coldspot contours}

\item{lwd}{a numeric, line width of the contours}

\item{pch}{an integer, the points type when \code{plotPoints=TRUE}.}

\item{point.size}{a numeric, the size of the points when \code{plotPoints=TRUE}.}

\item{alpha}{a fraction, alpha chanel for the raste points when \code{plotPoints=TRUE}.}

\item{label.cex}{a numeric, the size of the region label text.}

\item{xlim}{a numeric vector specifying \code{xlim} in msi coordinates.}

\item{ylim}{a numeric vector specifying \code{ylim} in msi coordinates.}

\item{...}{ignored.}
}
\value{
Generates plots and writes them to \code{outputPath}.
}
\description{
This functions plots MSI measurement locations (pixels) on top of the
corresponding (reference) optical image for visual inspection.
}
\author{
Denis Abu Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
