% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_SHAP_remover.R
\name{Zero_remover}
\alias{Zero_remover}
\title{Zero_SHAP_remover}
\usage{
Zero_remover(df, col_idx, ...)
}
\arguments{
\item{df}{a dataframe containing SHAP values for each tissue histology feature, this should contain the following vectors:
\enumerate{
\item "mzValues" as ranked by SHAP, 2) Average SHAP value for VT, 3) SHAP value for Necrosis, ... etc
}}

\item{col_idx}{a range of columns, specifying which tissue histology features to consider, for example 2:3 for VT and Nec.}

\item{...}{ignored.}
}
\value{
clean dataframe that filters SHAP values DataFrames.
}
\description{
a function to dynamically remove values from a dataframe based on a criterion
when a given number of columns have zero values in that particular list of columns.
It removes majority of irrelevant features in the m/z axis
}
\author{
Shad Arif Mohammed, \email{s.mohammed@doktoranden.hs-mannheim.de}
}
